fix pipeline for quickstart
