+++ src/app/space/forge-wizard/quickstart-pages/step2/project-info-step.component.html
@@ -2,8 +2,8 @@
   <div style="min-height: 450px" class="row">
     <div class="col-xs-12">
       <div class="form-horizontal">
-        <!--<single-selection-dropdown class="input-group width-100" [field]="gui.inputs[0]" [form]="form">-->
-        <!--</single-selection-dropdown>-->
+        <single-selection-dropdown class="input-group width-100" [field]="gui.inputs[0]">
+        </single-selection-dropdown>
         <single-input class="input-group width-100" [field]="gui.inputs[1]" [form]="form">
         </single-input>
         <single-input class="input-group width-100" [field]="gui.inputs[2]" [form]="form">
