+++ src/app/space/forge-wizard/quickstart-pages/step3/pipeline-quickstart-step.component.html
@@ -1,0 +1,14 @@
+<section *ngIf="gui.inputs">
+  <div style="min-height: 450px" class="row">
+    <div class="col-xs-12">
+      <div class="form-horizontal">
+        <label title="pipeline" required>Select a build pipeline strategy</label>
+        <pipeline-view
+          class="input-group width-100" [field]="gui.inputs[0]">
+        </pipeline-view>
+        <forge-errors [field]="gui"></forge-errors>
+      </div>
+    </div>
+  </div>
+</section>
+
