+++ src/app/space/forge-wizard/quickstart-pages/step3/pipeline-quickstart-step.component.ts
@@ -1,0 +1,23 @@
+import {Component, OnInit, Input} from '@angular/core';
+import { Gui } from '../../gui.model';
+import { FormGroup } from '@angular/forms';
+
+@Component({
+  selector: 'pipeline-quickstart-step',
+  templateUrl: './pipeline-quickstart-step.component.html'
+})
+export class PipelineQuickstartStepComponent implements OnInit {
+
+  @Input() gui: Gui;
+  @Input() form: FormGroup;
+  @Input() labelSpace: string;
+  constructor() {}
+
+  ngOnInit(): void {
+    // When creating pipeline we need to "tag" them the space label in OSO
+    // so that OSiO can filtered by space
+    this.gui.inputs[2].value = this.labelSpace;
+    console.log(`::::PipelineComponent ngOnInit: ${JSON.stringify(this.gui)}`);
+  }
+
+}
