+++ src/app/space/forge-wizard/quickstart-wizard.component.html
@@ -16,13 +16,11 @@
   <div [formGroup]="form" [ngSwitch]="currentGui.metadata?.name">
     <pfng-wizard-step [config]="steps[0]">
       <pfng-wizard-substep [config]="steps[3]">
-        booooooo {{currentGui.metadata?.name}}
         <span *ngSwitchCase="'Quickstart'">
           <quickstart [gui]="currentGui" [form]="form"></quickstart>
         </span>
       </pfng-wizard-substep>
       <pfng-wizard-substep [config]="steps[4]">
-        booooooo {{currentGui.metadata?.name}}
         <span *ngSwitchCase="'Fabric8: Project Info'">
           <project-info-step [gui]= "currentGui" [form]="form"></project-info-step>
         </span>
@@ -31,16 +29,12 @@
 
 
     <pfng-wizard-step [config]="steps[1]">
-
       <pfng-wizard-substep [config]="steps[5]">
-        booooooo {{currentGui.metadata?.name}}
         <span *ngSwitchCase="'Obsidian: Configure Pipeline'">
-          <pipeline-step [gui]="currentGui" [form]="form" [labelSpace]="currentSpace.attributes.name"></pipeline-step>
+          <pipeline-quickstart-step [gui]="currentGui" [form]="form" [labelSpace]="currentSpace.attributes.name"></pipeline-quickstart-step>
         </span>
       </pfng-wizard-substep>
-
       <pfng-wizard-substep [config]="steps[6]">
-        booooooo {{currentGui.metadata?.name}}
         <span *ngSwitchCase="'io.fabric8.forge.generator.kubernetes.CreateBuildConfigStep'">
           <build-config-step [gui]="currentGui" [form]="form"></build-config-step>
         </span>
@@ -49,7 +43,6 @@
 
     <pfng-wizard-step [config]="steps[2]">
       <pfng-wizard-substep [config]="steps[7]">
-        booooooo {{currentGui.metadata?.name}}
         <span *ngSwitchCase="'Review'">
           <review-step [gui]="currentGui"></review-step>
         </span>
